refactor: separate gamestate components
