+++ backend/gamestate/gamestate.py
@@ -3,14 +3,15 @@
 
 from enum import Enum
 from typing import List
-import json
 import os
 import time
 import logging
-from models import Team, Question
 from buzz_interface import Buzz
 from requests.exceptions import ConnectionError as requests_ConnectionError
-
+from .questions_controller import QuestionsController
+from .teams_controller import TeamsController
+from .actions import Actions
+from .models import Team, Question
 
 SPLIT_OR_STEAL = bool(os.getenv("USE_SPLIT_OR_STEAL", "True"))
 BUZZ_PENALTY_TIMEOUT = int(os.getenv("BUZZ_PENALTY_TIMEOUT", "5")) * 1e9
@@ -27,341 +28,6 @@ class States(Enum):
     TEAM_SELECTED = 4
     SPLIT_OR_STEAL = 5
     OVER = 6
-
-
-class QuestionsController:
-    """Class for controling questions atributes withtin the game"""
-
-    def __init_questions(self):
-        """Generate questions from questions.json
-
-        Raises:
-            ValueError: An even number of tiebreaker questions were given in the json
-        """
-        idx = 0
-        with open("backend/questions.json", "r", encoding="utf-8") as f:
-            data = json.load(f)
-
-            for category, questions in data["regular"].items():
-                for question in questions:
-                    self.questions.append(
-                        Question(
-                            idx,
-                            question["question"],
-                            question["answer"],
-                            question["image"],
-                            question["value"],
-                            category,
-                        )
-                    )
-                    idx += 1
-            for question in data["tiebreaker"]:
-                self.tiebreak_questions.append(
-                    Question(
-                        idx,
-                        question["question"],
-                        question["answer"],
-                        question["image"],
-                        100,
-                        "Tiebreak",
-                        True,
-                    )
-                )
-                idx += 1
-            if (
-                len(self.tiebreak_questions) % 2 == 0
-            ):  # Make sure there are an odd number of tiebreak questions
-                raise ValueError("There must be an odd number of tiebreak questions")
-
-            self.current_question_idx = 0
-
-    def __init__(self):
-        self.questions: List[Question] = []
-        self.current_question_idx: int = None
-        self.tiebreak_questions: List[Question] = []
-        self.in_tiebreak = False
-        self.__init_questions()
-
-    def get_current_question(self) -> Question:
-        """
-            Fetches the question being played
-
-        Returns:
-            Question: The current question
-        """
-        if self.in_tiebreak:
-            return self.tiebreak_questions[self.current_question_idx]
-        return self.questions[self.current_question_idx]
-
-    def select_question(self, idx: int = None):
-        """Sets a new question as the one being played
-
-        Args:
-            idx (int, optional): the index of the question aka its id. Defaults to None.
-
-        Raises:
-            ValueError: idx is invalid
-        """
-        if idx is None and not self.in_tiebreak:
-            raise ValueError("No question selected")
-        if self.in_tiebreak:
-            self.current_question_idx += 1
-        else:
-            if idx < 0 or idx >= len(self.questions):
-                raise ValueError("Index out of bonds")
-            self.current_question_idx = idx
-
-    def questions_over(self) -> bool:
-        """checks if all questions have benn answered
-
-        Returns:
-            bool: are all questions answered
-        """
-        return all(map(lambda q: q.answered, self.questions))
-
-    def skip(self):
-        """
-        Skips the current question
-        """
-        self.questions[self.current_question_idx].skip()
-
-    def tiebreak(self):
-        """
-        makes the game be in a tiebreak state
-        if already in a tiebreak state selects the next question
-        """
-        if not self.in_tiebreak:
-            self.current_question_idx = 0
-            self.in_tiebreak = True
-        else:
-            self.current_question_idx += 1
-
-    def __answer_correctly(self, team: Team):
-        """action for a team answering a question correctly
-
-        Args:
-            team (Team): team that answered the question
-        """
-        self.questions[self.current_question_idx].answer_correctly(team)
-
-    def __answer_incorreclty(self, team: Team):
-        """action for a team answering a question incorrectly
-
-        Args:
-            team (Team): team that answered the question
-        """
-        self.questions[self.current_question_idx].answer_incorreclty(team)
-
-    def answer(self, team: Team, correct: bool):
-        """action for answering a question
-
-        Args:
-            team (Team): team that answered
-            correct (bool): the answer is correct
-        """
-        if correct:
-            self.__answer_correctly(team)
-        else:
-            self.__answer_incorreclty(team)
-
-    def list_questions(self) -> List[Question]:
-        """lists all question in the game
-
-        Returns:
-            List[Question]: list of all question in the game
-        """
-        return self.questions.copy()
-
-    def get_question(self, idx: int) -> Question:
-        """gets a question by id
-
-        Args:
-            idx (int): the index of the question
-
-        Raises:
-            ValueError: idx is none
-            ValueError: idx is invalid
-        """
-        if idx is None or idx < 0 or idx >= len(self.questions):
-            return None
-        return self.questions[idx]
-
-
-class TeamsController:
-    """Class for controling teams atributes withtin the game"""
-
-    def __init__(self):
-        self.teams: List[Team] = []
-        self.current_playing_id: int = 0
-        self.selecting_id: int = 0
-        self.playing: List[int] = []
-
-    def get_current_team(self) -> Team:
-        """get the team playing
-
-        Returns:
-            Team: the team playing
-        """
-        if self.teams == []:
-            return None
-        return self.teams[self.current_playing_id]
-
-    def get_team(self, idx: int) -> Team:
-        """get team with given id
-
-        Args:
-            idx (int): id of team
-
-        Returns:
-            Team: team with given id
-        """
-        if self.teams == []:
-            return None
-        return self.teams[idx]
-
-    def get_selecting_team(self) -> Team:
-        """get the team selecting
-
-        Returns:
-            Team: the team selecting
-        """
-        if self.teams == []:
-            return None
-        return self.teams[self.selecting_id]
-
-    def set_teams(self, player_teams_names: List[List[str]]):
-        """creates the teams with the given names
-
-        Args:
-            teams_names (List[str]): list of names
-        Raises:
-            ValueError: Too many teams (max 4)
-            ValueError: Too many players in a team (max 4)
-        """
-        if len(player_teams_names) > 4:
-            raise ValueError("Too many teams (max 4)")
-        for t in player_teams_names:
-            if len(t) > 4:
-                raise ValueError("Too many players in a team (max 4)")
-        self.teams = [Team(idx, names) for idx, names in enumerate(player_teams_names)]
-        self.current_playing_id = 0
-        self.selecting_id = 0
-        self.playing = list(range(len(player_teams_names)))
-
-    def split_or_steal(self, votes: List[int]):
-        """split or steal the balance of the team
-
-        Args:
-            members (List[int]): list of members that stole.
-        """
-        logging.debug("Split or Steal: [%s]", ",".join(str(v) for v in votes))
-        team = self.get_winning_team()
-        if len(votes) > 1:
-            team.balance = 0
-        elif len(votes) == 1:
-            stealer = votes[0]
-            new_id = len(self.teams)
-            self.teams.append(Team(new_id, [team.names[stealer]]))
-            self.teams[new_id].balance = team.balance
-            team.balance = 0
-            del team.names[stealer]
-
-    def set_selecting(self, idx: int):
-        """set team with given id as selecting
-
-        Args:
-            idx (int): id of team to set as selecting
-        """
-        self.selecting_id = idx
-
-    def set_current_playing(self, idx: int):
-        """set team with given id as playing
-
-        Args:
-            idx (int): id of team to set as playing
-        """
-        self.current_playing_id = idx
-
-    def is_tie(self) -> bool:
-        """check if the teams are in a tie
-
-        Returns:
-            bool: the teams are in a tie
-        """
-        max_tokens = max(p.balance for p in self.teams)
-        teams_with_max_tokens = [p.id for p in self.teams if p.balance == max_tokens]
-        if len(teams_with_max_tokens) > 1:
-            self.playing = teams_with_max_tokens
-            return True
-        return False
-
-    def get_winning_team(self) -> Team:
-        """get the winning team
-
-        Returns:
-            Team: the winning team
-        """
-        if self.teams == []:
-            return None
-        c = self.teams[0]
-        for t in self.teams[1:]:
-            if t.balance > c.balance:
-                c = t
-        return c
-
-    def next_selecting(self):
-        """make the next team in line be the selecting one"""
-        self.selecting_id = (self.selecting_id + 1) % 4
-
-    def set_selecting_as_current(self):
-        """make the team playing as the selecting"""
-        self.selecting_id = self.current_playing_id
-
-    def list_teams(self) -> List[Team]:
-        """return the list of the teams in the game
-
-        Returns:
-            List[Team]: list of teams in the game
-        """
-        return self.teams.copy()
-
-
-class Actions:
-    """Class for controling actions atributes withtin the game"""
-
-    def __init__(self):
-        self.play_correct_sound: bool = False
-        self.play_wrong_sound: bool = False
-        self.play_start_accepting: bool = False
-        self.play_buzzer_sound: bool = False
-        self.stop_countdown: bool = False
-
-    def reset_sound(self):
-        """
-        Set both sound playing to false
-        """
-        self.play_correct_sound = False
-        self.play_wrong_sound = False
-        self.play_start_accepting = False
-        self.play_buzzer_sound = False
-
-    def reset_countdown_timer(self):
-        """reset the countdown timer"""
-        self.stop_countdown = False
-
-    def to_dict(self) -> dict:
-        """
-            cast the state as a dictionary to represent as json object
-        Returns:
-            dict: the gamestate data
-        """
-        return {
-            "playCorrectSound": self.play_correct_sound,
-            "playWrongSound": self.play_wrong_sound,
-            "playStartAccepting": self.play_start_accepting,
-            "playBuzzerSound": self.play_buzzer_sound,
-            "stopTimer": self.stop_countdown,
-        }
 
 
 class GameState:
