--- backend/shared_globals.py
@@ -1,5 +1,0 @@
-"""Module for storing globals shared by all the threads"""
-
-from gamestate import GameState
-
-state: GameState = None
